+++ shared/src/main/java/net/onebeastchris/hurricane/config/HurricaneConfiguration.java
@@ -11,11 +11,6 @@ public final class HurricaneConfiguration {
             Caveats: a custom client - Java or Bedrock - could take advantage of no collision and walk right through.
              Additionally, placement of these blocks on both platforms may be buggier than usual.""")
     private CollisionFixes collisionFixes = new CollisionFixes();
-    @Comment("""
-            Fixes Bedrock players being unable to control pigs and striders by controlling their movement serverside.
-            Java Edition controls pigs and striders on the client end. Bedrock depends on the server.
-            This option should be relatively safe but does modify server behavior. Geyser or Floodgate must be installed.""")
-    private boolean itemSteerableFix = true;
 
     @Comment("""
             Suppresses "Mismatch in destroy block pos" warnings so they don't spam each time e.g. grass is broken.\s
@@ -26,10 +21,6 @@ public final class HurricaneConfiguration {
     @SuppressWarnings("ClassEscapesDefinedScope")
     public CollisionFixes collisionFixes() {
         return collisionFixes;
-    }
-
-    public boolean itemSteerableFix() {
-        return itemSteerableFix;
     }
 
     public boolean suppressWarnings() {