+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-Hurricane 2.0:
-- Added 1.21 support
-- Added NeoForge port
+Hurricane 2.1:
+- Added config version
+- Removed the item steerables workaround due to it being implemented natively in Geyser